Control.Concurrent.Thread.Utils

Description

This module provides various helpful utilities for dealing with threads.

Written by John Goerzen, jgoerzen@complete.org

I/O utilities

runInThread :: IO a -> (a -> IO b) -> IO ThreadId

Takes a IO action and a function. The IO action will be called in a separate thread. When it is completed, the specified function is called with its result. This is a simple way of doing callbacks.