Data.Bits.Utils

Description

Bit-related utilities

Written by John Goerzen, jgoerzen@complete.org

Documentation

getBytes :: (Integral a, Bounded a, Bits a) => a -> [a]

Returns a list representing the bytes that comprise a data type.

Example:

getBytes (0x12345678::Int) -> [0x12, 0x34, 0x56, 0x78]

fromBytes :: (Bits a, Num a) => [a] -> a

The opposite of getBytes, this function builds a number based on its component bytes.

Results are undefined if any components of the input list are > 0xff!

c2w8 :: Char -> Word8

Converts a Char to a Word8.

s2w8 :: String -> [Word8]

Converts a String to a [Word8].

w82c :: Word8 -> Char

Converts a Word8 to a Char.

w82s :: [Word8] -> String

Converts a [Word8] to a String.