Data.Maybe.Utils

Description

Utilities for working with the Either data type

Documentation

forceMaybe :: Maybe a -> a

Pulls a Just value out of a Maybe value. If the Maybe value is Nothing, raises an exception with error.

forceMaybeMsg :: String -> Maybe a -> a

Like forceMaybe, but lets you customize the error message raised if Nothing is supplied.