System.IO.HVFS.Combinators

Description

Support for combining different HVFS modules together

Copyright (c) 2004-2005 John Goerzen, jgoerzen@complete.org

Restrictions

data HVFS a => HVFSReadOnly a

Restrict access to the underlying filesystem to be strictly read-only. Any write-type operations will cause an error.

No constructor is required; just say HVFSReadOnly fs to make a new read-only wrapper around the HVFS instance fs.

Constructors

HVFSReadOnly a

data HVFS a => HVFSChroot a

Access a subdirectory of a real filesystem as if it was the root of that filesystem.

newHVFSChroot

Arguments

:: HVFS a

=> a
The object to pass requests on to

-> FilePath
The path of the directory to make root

-> IO (HVFSChroot a)
The resulting new object

Create a new HVFSChroot object.