System.Debian

Description

This module provides various helpful utilities for dealing with Debian files and programs.

Written by John Goerzen, jgoerzen@complete.org

Control or Similar File Utilities

type ControlFile = [(String, String)]

The type representing the contents of a Debian control file, or any control-like file (such as the output from apt-cache show, etc.)

Version Number Utilities

data DebVersion

The type representing a Debian version number. This type is an instance of Ord, but you can also use compareDebVersion if you prefer.

compareDebVersion :: String -> String -> IO Ordering

Compare the versions of two packages.

checkDebVersion

Arguments

:: String
Version 1

-> String
Operator

-> String
Version 2

-> IO Bool