System.IO.HVFS.Utils

Description

This module provides various helpful utilities for dealing filesystems.

Written by John Goerzen, jgoerzen@complete.org

To operate on your system's main filesystem, just pass SystemFS as the first parameter to these functions.

Documentation

recurseDir :: HVFS a => a -> FilePath -> IO [FilePath]

Obtain a recursive listing of all files/directories beneath the specified directory. The traversal is depth-first and the original item is always present in the returned list.

If the passed value is not a directory, the return value be only that value.

The "." and ".." entries are removed from the data returned.

recurseDirStat :: HVFS a => a -> FilePath -> IO [(FilePath, HVFSStatEncap)]

Like recurseDir, but return the stat() (System.Posix.Files.FileStatus) information with them. This is an optimization if you will be statting files yourself later.

The items are returned lazily.

WARNING: do not change your current working directory until you have consumed all the items. Doing so could cause strange effects.

Alternatively, you may wish to pass an absolute path to this function.

recursiveRemove :: HVFS a => a -> FilePath -> IO ()

Removes a file or a directory. If a directory, also removes all its child files/directories.

lsl :: HVFS a => a -> FilePath -> IO String

Provide a result similar to the command ls -l over a directory.

Known bug: setuid bit semantics are inexact compared with standard ls.

data SystemFS

Constructors

SystemFS