System.Path.Glob

Description

Functions for expanding wildcards, filenames, and pathnames.

For information on the metacharacters recognized, please see the notes in System.Path.WildMatch.

Documentation

glob :: FilePath -> IO [FilePath]

Takes a pattern. Returns a list of names that match that pattern. The pattern is evaluated by System.Path.WildMatch. This function does not perform tilde or environment variable expansion.

Filenames that begin with a dot are not included in the result set unless that component of the pattern also begins with a dot.

In MissingH, this function is defined as:

glob = vGlob SystemFS

vGlob :: HVFS a => a -> FilePath -> IO [FilePath]

Like glob, but works on both the system ("real") and HVFS virtual filesystems.